# Improve perf of uniform setters