--- a/test/webgl/uniforms.spec.js
+++ b/test/webgl/uniforms.spec.js
@@ -39,13 +39,20 @@ precision highp float;
 #endif
 
 uniform float f;
-uniform int i;
-uniform bool b;
 uniform vec2 v2;
 uniform vec3 v3;
 uniform vec4 v4;
-// int vectors
-// bool vectors
+
+uniform int i;
+uniform ivec2 iv2;
+uniform ivec3 iv3;
+uniform ivec4 iv4;
+
+uniform bool b;
+uniform bvec2 bv2;
+uniform bvec3 bv3;
+uniform bvec4 bv4;
+
 uniform mat2 m2;
 uniform mat3 m3;
 uniform mat4 m4;
@@ -54,24 +61,40 @@ uniform sampler2D s2d;
 // uniform samplerCube sCube;
 
 void main(void) {
+  vec4 v = vec4(f) + vec4(v2, 0., 0.) + vec4(v3, 0.) + v4;
+  ivec4 iv = ivec4(i) + ivec4(iv2, 0., 0.) + ivec4(iv3, 0.) + iv4;
+
+  bvec4 bv = bv4;
+  bv = bvec4(bv3, 0.);
+  bv = bvec4(bv2, 0., 0.);
+  bv = bvec4(b);
+
+  vec2 transform_v2 = m2 * v2;
+  vec3 transform_v3 = m3 * v3;
+  vec4 transform_v4 = m4 * v4;
+
+  v = texture2D(s2d, v2);
+
   gl_FragColor = vec4(1.0, 1.0, 1.0, 1.0);
 }
 `;
 
 const WEBGL1_GOOD_UNIFORMS = {
   f: 1.0,
-  i: 1,
-  b: true,
   v2: new Float32Array([1, 2]), // FLOAT_VEC2  0x8B50
   v3: new Float32Array([1, 2, 3]), // FLOAT_VEC3  0x8B51
   v4: new Float32Array([1, 2, 3, 4]), // FLOAT_VEC4  0x8B52
-  // INT_VEC2  0x8B53
-  // INT_VEC3  0x8B54
-  // INT_VEC4  0x8B55
-  // BOOL  0x8B56
-  // BOOL_VEC2 0x8B57
-  // BOOL_VEC3 0x8B58
-  // BOOL_VEC4 0x8B59
+
+  i: -1,
+  iv2: new Int32Array([1, 2]), // INT_VEC2  0x8B53
+  iv3: new Int32Array([1, 2, 3]), // INT_VEC3  0x8B54
+  iv4: new Int32Array([1, 2, 3, 4]), // INT_VEC4  0x8B55
+
+  b: true, // BOOL  0x8B56
+  bv2: new Int32Array([false, true]), // BOOL_VEC2 0x8B57
+  bv3: new Int32Array([false, true, false]), // BOOL_VEC3 0x8B58
+  bv4: new Int32Array([false, true, false, true]), // BOOL_VEC4 0x8B59
+
   m2: new Float32Array(MATRIX_2), // FLOAT_MAT2  0x8B5A
   m3: new Float32Array(MATRIX_3), // FLOAT_MAT3  0x8B5B
   m4: new Float32Array(MATRIX_4), // FLOAT_MAT4  0x8B5C
@@ -155,14 +178,72 @@ test('WebGL#Uniforms pre verify uniforms', t => {
   t.end();
 });
 
-test('WebGL#Uniforms Program construct/delete', t => {
+test('WebGL#Uniforms Program uniform locations', t => {
   const {gl} = fixture;
 
   const program = new Program(gl, {
     vs: VERTEX_SHADER,
     fs: WEBGL1_FRAGMENT_SHADER
   });
-  t.ok(program instanceof Program, 'Program construction successful');
+
+  for (const uniformName in WEBGL1_GOOD_UNIFORMS) {
+    t.ok(program._uniformSetters[uniformName], `Program found uniform setter ${uniformName}`);
+  }
 
   t.end();
+});
+
+const testSetUniform = gl => t => {
+  const program = new Program(gl, {
+    vs: VERTEX_SHADER,
+    fs: WEBGL1_FRAGMENT_SHADER
+  });
+  program.use();
+
+  const uniforms = Object.assign({}, WEBGL1_GOOD_UNIFORMS);
+
+  program.setUniforms(uniforms);
+  t.pass('Program set uniforms successful');
+
+  for (const uniformName in WEBGL1_GOOD_UNIFORMS) {
+    const value = WEBGL1_GOOD_UNIFORMS[uniformName];
+    if (value.length) {
+      // Convert to plain array
+      uniforms[uniformName] = Array.from(value);
+    }
+  }
+
+  program.setUniforms(uniforms);
+  t.pass('Program set array uniforms successful');
+
+  for (const uniformName in WEBGL1_GOOD_UNIFORMS) {
+    const value = WEBGL1_GOOD_UNIFORMS[uniformName];
+    if (value.length) {
+      // Convert to wrong typed array
+      uniforms[uniformName] = (value instanceof Float32Array) ?
+        new Int32Array(value) : new Float32Array(value);
+    }
+  }
+
+  program.setUniforms(uniforms);
+  t.pass('Program set malformed uniforms successful');
+
+  t.end();
+};
+
+test('WebGL#Uniforms Program setUniforms', t => {
+  const {gl} = fixture;
+
+  testSetUniform(gl)(t);
+});
+
+test('WebGL2#Uniforms Program setUniforms', t => {
+  const {gl2} = fixture;
+
+  if (gl2) {
+    testSetUniform(gl2)(t);
+  } else {
+    t.end();
+  }
+
 });