The methods accept either a Float32Array or an array of numbers: [uniformMatrix](https://developer.mozilla.org/en-US/docs/Web/API/WebGLRenderingContext/uniformMatrix), [uniform](https://developer.mozilla.org/en-US/docs/Web/API/WebGLRenderingContext/uniform)

Node / WebGL1:

| Test | Result |
| ----- | ------ |
| Set vec3 uniform from Float32Array | 870k iterations/s (0.12s) |
| Set vec3 uniform from cloned Float32Array | 467k iterations/s (0.21s) |
| Set vec3 uniform from array | 794k iterations/s (0.13s) |
| Set vec3 uniform from new Float32Array | 505k iterations/s (0.20s) |
| Set mat4 uniform from Float32Array | 422k iterations/s (0.24s) |
| Set mat4 uniform from cloned Float32Array | 311k iterations/s (0.32s) |
| Set mat4 uniform from array | 389k iterations/s (0.26s) |
| Set mat4 uniform from new Float32Array | 296k iterations/s (0.34s) |


Browser / WebGL2 (Chrome 62):

| Test | Result |
| ----- | ------ |
| Set vec3 uniform from Float32Array | 2.10M iterations/s (0.48s) |
| Set vec3 uniform from cloned Float32Array | 1.88M iterations/s (0.53s) |
| Set vec3 uniform from array | 1.74M iterations/s (0.58s) |
| Set vec3 uniform from new Float32Array | 521k iterations/s (0.19s) |
| Set mat4 uniform from Float32Array | 1.90M iterations/s (0.53s) |
| Set mat4 uniform from cloned Float32Array | 339k iterations/s (0.29s) |
| Set mat4 uniform from array | 485k iterations/s (0.21s) |
| Set mat4 uniform from new Float32Array | 170k iterations/s (0.59s) |

Tested: node, browser, website