@Pessimistress 2017-12-07
@ibgreen Thanks for the pre-allocated arrays suggestion, it actually works better than plain array, so I will just cast everything to typed array.
After the change:

Node / WebGL1

| Test | Result |
| ----- | ------ |
| Set vec3 uniform from Float32Array | 685k iterations/s (0.15s) |
| Set vec3 uniform from plain array | 787k iterations/s (0.13s) |
| Set vec3 uniform from Int32Array | 662k iterations/s (0.15s) |
| Set mat4 uniform from Float32Array | 472k iterations/s (0.21s) |
| Set mat4 uniform from plain array | 372k iterations/s (0.27s) |
| Set mat4 uniform from Int32Array | 392k iterations/s (0.26s) |

Node / WebGL2 (Chrome 62)

| Test | Result |
| ----- | ------ |
| Set vec3 uniform from Float32Array | 2.75M iterations/s (0.36s) |
| Set vec3 uniform from plain array | 2.86M iterations/s (0.35s) |
| Set vec3 uniform from Int32Array | 2.64M iterations/s (0.38s) |
| Set mat4 uniform from Float32Array | 2.06M iterations/s (0.48s) |
| Set mat4 uniform from plain array | 1.17M iterations/s (0.85s) |
| Set mat4 uniform from Int32Array | 1.00M iterations/s (0.10s) |